requirements.txt
streamlit==1.45.1
langchain==0.1.17
langchain-community==0.0.36
chromadb==0.4.24
sentence-transformers==2.2.2
huggingface-hub==0.23.1
python-dotenv==1.0.1